Added _job view and rendered job view for show page
Used jobs.create to save new jobs. These are rendered through jobs
view, and are now visible on student show page. Need to allow for
editing and deleting. Also need to make method for some of the long
descriptors of jobs so that they don’t have to be re-typed each time.
Also, contacts.